Refactor code to services an react query
